pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "foundrygate"
version = "0.3.0"
description = "Local OpenAI-compatible routing gateway for OpenClaw and other AI-native clients."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["foundrygate", "openclaw", "llm", "router", "gateway", "proxy", "multi-provider"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Internet :: Proxy Servers",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx",  # for TestClient
    "ruff>=0.8",
]

[project.scripts]
foundrygate = "foundrygate.main:main"
foundrygate-stats = "foundrygate.cli:main"

[project.urls]
Homepage = "https://github.com/typelicious/ClawGate"
Repository = "https://github.com/typelicious/ClawGate"
Issues = "https://github.com/typelicious/ClawGate/issues"
Documentation = "https://github.com/typelicious/ClawGate#readme"

[tool.setuptools.packages.find]
include = ["foundrygate*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
